python/requirements.txt
pyshp
mapbox
python-dotenv
